+++ src/input/source/tty.rs
@@ -5,10 +5,10 @@ use mio::{unix::EventedFd, Events, Poll, PollOpt, Ready, Token};
 use crate::Result;
 
 use super::super::{
-    event_source::EventSource,
     events::InternalEvent,
-    poll_timeout::PollTimeout,
+    source::EventSource,
     sys::unix::{parse_event, tty_fd, FileDesc},
+    timeout::PollTimeout,
 };
 
 // Tokens to identify file descriptor
